feat: optimize polling worker
