+++ src/AeFinder.BackgroundWorker.Core/Options/ScheduledTaskOptions.cs
@@ -13,6 +13,8 @@ public class ScheduledTaskOptions
     public int RenewalAdvanceWarningDays { get; set; } = 5;
     public int BillingIndexerPollingTaskPeriodMilliSeconds { get; set; } = 10000;
     public int OrderPaymentResultPollingTaskPeriodMilliSeconds { get; set; } = 12000;
+    public int AdvanceBillPaymentResultPollingTaskPeriodMilliSeconds { get; set; } = 300000;
+    public int SettlementBillPaymentResultPollingTaskPeriodMilliSeconds { get; set; } = 180000;
     public int UnpaidOrderTimeoutMinutes { get; set; } = 30;
     public int UnpaidBillTimeOutDays { get; set; } = 7;
 }
